backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
PyMuPDF==1.27.1
beautifulsoup4==4.12.3
markdownify==0.13.1
